Moved comments state to Comments.js component. Replaced unnecessary image props with appropriate imports.
